src/requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.7
mistralai==0.0.9
requests==2.31.0
beautifulsoup4==4.12.2